[SPOT-286][REFACTOR] 잘못 분리된 story 코드 이동
