+++ src/main/java/com/example/spot/story/application/StoryQueryService.java
@@ -3,6 +3,8 @@ package com.example.spot.story.application;
 import com.example.spot.story.domain.enums.StoryCategoryQuery;
 import com.example.spot.story.web.dto.response.StoryCommentResponseDTO;
 import com.example.spot.story.web.dto.response.StoryResDTO;
+import com.example.spot.story.web.dto.response.StoryResponseDTO;
+import com.example.spot.study.presentation.dto.response.StudyImageResponseDTO;
 import org.springframework.data.domain.PageRequest;
 
 public interface StoryQueryService {
@@ -13,6 +15,13 @@ public interface StoryQueryService {
     // 스터디 게시글 불러오기
     StoryResDTO.PostDetailDTO getPost(Long studyId, Long postId, Boolean likeOrScrap);
 
+    // 스터디 공지 게시글 불러오기
+    StoryResponseDTO findStudyAnnouncementPost(Long studyId);
+
     // 스터디 게시글 댓글 목록 불러오기
     StoryCommentResponseDTO.CommentReplyListDTO getAllComments(Long studyId, Long postId);
+
+    // 스터디 이미지 목록 조회
+    StudyImageResponseDTO.ImageListDTO getAllStudyImages(Long studyId, PageRequest pageRequest);
+
 }
