+++ src/main/java/com/example/spot/story/web/controller/StoryController.java
@@ -3,9 +3,9 @@ package com.example.spot.story.web.controller;
 import com.example.spot.common.api.ApiResponse;
 import com.example.spot.common.api.code.status.SuccessStatus;
 import com.example.spot.story.domain.enums.StoryCategoryQuery;
-import com.example.spot.common.application.s3.S3ImageService;
 import com.example.spot.story.application.StoryCommandService;
 import com.example.spot.story.application.StoryQueryService;
+import com.example.spot.story.web.dto.response.StoryResponseDTO;
 import com.example.spot.study.domain.validation.annotation.ExistStudy;
 import com.example.spot.story.domain.validation.annotation.ExistStory;
 import com.example.spot.story.domain.validation.annotation.ExistStoryComment;
@@ -14,6 +14,7 @@ import com.example.spot.story.web.dto.request.StoryRequestDTO;
 import com.example.spot.story.web.dto.response.StoryCommentResponseDTO;
 import com.example.spot.story.web.dto.response.StoryResDTO;
 
+import com.example.spot.study.presentation.dto.response.StudyImageResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -33,7 +34,6 @@ public class StoryController {
 
     private final StoryQueryService storyQueryService;
     private final StoryCommandService storyCommandService;
-    private final S3ImageService s3ImageService;
 
     /* ----------------------------- 스터디 게시글 관련 API ------------------------------------- */
 
@@ -118,6 +118,17 @@ public class StoryController {
             @RequestParam Boolean likeOrScrap) {
         StoryResDTO.PostDetailDTO postDetailDTO = storyQueryService.getPost(studyId, postId, likeOrScrap);
         return ApiResponse.onSuccess(SuccessStatus._STUDY_POST_FOUND, postDetailDTO);
+    }
+
+    @Tag(name = "스터디 상세 정보")
+    @Operation(summary = "[스터디 상세 정보] 스터디 최근 공지 1개 불러오기", description = """ 
+            ## [스터디 상세 정보] 내 스터디 > 스터디 클릭, 로그인한 회원이 참여하는 특정 스터디의 최근 공지 1개를 불러옵니다.
+            study_post의 announced_at이 가장 최근인 공지 1개가 반환됩니다.
+            """)
+    @GetMapping("/studies/{studyId}/announce")
+    public ApiResponse<StoryResponseDTO> getRecentAnnouncement(@PathVariable @ExistStudy Long studyId) {
+        StoryResponseDTO storyResponseDTO = storyQueryService.findStudyAnnouncementPost(studyId);
+        return ApiResponse.onSuccess(SuccessStatus._STUDY_POST_FOUND, storyResponseDTO);
     }
 
     @Tag(name = "스터디 게시글")
@@ -284,6 +295,23 @@ public class StoryController {
             @PathVariable @ExistStory Long postId) {
         StoryCommentResponseDTO.CommentReplyListDTO commentReplyListDTO = storyQueryService.getAllComments(studyId, postId);
         return ApiResponse.onSuccess(SuccessStatus._STUDY_POST_COMMENT_FOUND, commentReplyListDTO);
+    }
+
+/* ----------------------------- 스터디 게시글 이미지 관련 API ------------------------------------- */
+
+    @Tag(name = "스터디 게시글 - 갤러리")
+    @Operation(summary = "[스터디 게시글 - 갤러리] 스터디 이미지 목록 불러오기", description = """ 
+            ## [스터디 게시글] 내 스터디 > 스터디 > 갤러리 클릭, 로그인한 회원이 참여하는 스터디의 이미지 목록을 불러옵니다.
+            스터디에 존재하는 모든 게시글의 이미지를 최신순으로 반환합니다.
+            """)
+    @Parameter(name = "studyId", description = "이미지 목록을 불러올 스터디의 id를 입력합니다.", required = true)
+    @GetMapping("/studies/{studyId}/images")
+    public ApiResponse<StudyImageResponseDTO.ImageListDTO> getAllStudyImages(
+            @PathVariable @ExistStudy Long studyId,
+            @RequestParam @Min(0) Integer offset,
+            @RequestParam @Min(1) Integer limit) {
+        StudyImageResponseDTO.ImageListDTO imageListDTO = storyQueryService.getAllStudyImages(studyId, PageRequest.of(offset, limit));
+        return ApiResponse.onSuccess(SuccessStatus._STUDY_POST_IMAGES_FOUND, imageListDTO);
     }
 
 }
