+++ src/main/java/com/example/spot/study/application/StudyCommandService.java
@@ -8,10 +8,11 @@ import com.example.spot.study.presentation.dto.response.StudyRegisterResponseDTO
 
 public interface StudyCommandService {
 
+    // 스터디 참여 신청하기
     StudyJoinResponseDTO.JoinDTO applyToStudy(Long studyId, StudyJoinRequestDTO.StudyJoinDTO studyJoinRequestDTO);
 
+    // 스터디 등록하기
     StudyRegisterResponseDTO.RegisterDTO registerStudy(StudyRegisterRequestDTO.RegisterDTO studyRegisterRequestDTO);
-
 
     // TODO 스터디 좋아요 Member 도메인 하위로 옮기기
     StudyLikeResponseDTO likeStudy(Long memberId, Long studyId);
