+++ src/main/java/com/example/spot/study/application/StudyMemberCommandService.java
@@ -16,14 +16,19 @@ import jakarta.validation.Valid;
 
 public interface StudyMemberCommandService {
 
+    // 스터디 탈퇴
     StudyWithdrawalResponseDTO.WithdrawalDTO withdrawFromStudy(Long studyId);
+
+    // 스터디 호스트 탈퇴
     StudyWithdrawalResponseDTO.WithdrawalDTO withdrawHostFromStudy(Long studyId, StudyHostWithdrawRequestDTO requestDTO);
 
+    // 스터디 종료
     StudyTerminationResponseDTO.TerminationDTO terminateStudy(Long studyId, String performance);
 
     // 스터디 신청 수락
     StudyApplyResponseDTO acceptAndRejectStudyApply(Long memberId, Long studyId, boolean isAccept);
 
+    // 스터디 승인 거절 테스트
     StudyApplyResponseDTO acceptAndRejectStudyApplyForTest(Long memberId, Long studyId, boolean isAccept);
 
     // 스터디 회원 신고
